[mod/#107] Int -> Long 변경

+++ data/src/main/java/com/ssing/data/instructorlessondetail/remote/datasource/api/InstructorLessonDetailDataSource.kt
@@ -5,6 +5,6 @@ import com.ssing.data.instructorlessondetail.remote.dto.response.InstructorLesso
 
 internal interface InstructorLessonDetailDataSource {
     suspend fun instructorLessonDetail(
-        lessonId: Int,
+        lessonId: Long,
     ): BaseResponse<InstructorLessonDetailResponse>
 }
