+++ data/src/main/java/com/ssing/data/instructorlessondetail/repository/api/InstructorLessonDetailRepository.kt
@@ -1,5 +1,5 @@
 package com.ssing.data.instructorlessondetail.repository.api
 
 interface InstructorLessonDetailRepository {
-    suspend fun instructorLessonDetail(lessonId: Int): Result<Unit>
+    suspend fun instructorLessonDetail(lessonId: Long): Result<Unit>
 }
